Add ScheduleClassesBase component
